requirements.txt
openai-whisper==20230314
fastapi==0.95.1
setuptools-rust==1.6.0
websockets==11.0.3
uvicorn==0.22.0